+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/datahub_client.py
@@ -386,3 +386,7 @@ class DataHubCatalogueClient(BaseCatalogueClient):
         return self.search_client.list_data_product_assets(
             urn=urn, count=count, start=start
         )
+
+    def get_glossary_terms(self, count: int = 1000) -> SearchResponse:
+        """Wraps the client's glossary terms query"""
+        return self.search_client.get_glossary_terms(count)