+++ python-libraries/data-platform-catalogue/tests/test_datahub_search.py
@@ -740,3 +740,77 @@ def test_list_data_product_assets(mock_graph, searcher):
             )
         ],
     )
+
+
+def test_get_glossary_terms(mock_graph, searcher):
+    datahub_response = {
+        "searchAcrossEntities": {
+            "start": 0,
+            "count": 2,
+            "total": 2,
+            "searchResults": [
+                {
+                    "entity": {
+                        "urn": "urn:li:glossaryTerm:022b9b68-c211-47ae-aef0-2db13acfeca8",
+                        "properties": {
+                            "name": "IAO",
+                            "description": "Information asset owner.\n",
+                        },
+                        "parentNodes": {
+                            "nodes": [
+                                {
+                                    "properties": {
+                                        "name": "Data protection terms",
+                                        "description": "Data protection terms",
+                                    }
+                                }
+                            ]
+                        },
+                    }
+                },
+                {
+                    "entity": {
+                        "urn": "urn:li:glossaryTerm:0eb7af28-62b4-4149-a6fa-72a8f1fea1e6",
+                        "properties": {
+                            "name": "Security classification",
+                            "description": "Only data that is 'official'",
+                        },
+                        "parentNodes": {"nodes": []},
+                    }
+                },
+            ],
+        }
+    }
+
+    mock_graph.execute_graphql = MagicMock(return_value=datahub_response)
+
+    response = searcher.get_glossary_terms(count=2)
+    print(response)
+    assert response == SearchResponse(
+        total_results=2,
+        page_results=[
+            SearchResult(
+                id="urn:li:glossaryTerm:022b9b68-c211-47ae-aef0-2db13acfeca8",
+                name="IAO",
+                description="Information asset owner.\n",
+                metadata={
+                    "parentNodes": [
+                        {
+                            "properties": {
+                                "name": "Data protection terms",
+                                "description": "Data protection terms",
+                            }
+                        }
+                    ]
+                },
+                result_type=ResultType.GLOSSARY_TERM,
+            ),
+            SearchResult(
+                id="urn:li:glossaryTerm:0eb7af28-62b4-4149-a6fa-72a8f1fea1e6",
+                name="Security classification",
+                description="Only data that is 'official'",
+                metadata={"parentNodes": []},
+                result_type=ResultType.GLOSSARY_TERM,
+            ),
+        ],
+    )